monitor/requirements.txt
asgiref==3.7.2
bidict==0.22.1
blinker==1.7.0
click==8.1.7
Flask==3.0.1
h11==0.14.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
python-dotenv==1.0.1
python-engineio==4.8.2
python-socketio==5.11.0
simple-websocket==1.0.0
waitress==2.1.2
websocket-client==1.7.0
Werkzeug==3.0.1
wsproto==1.2.0